requirements.txt
fastapi[all]
alembic
sqlalchemy
asyncpg
python-dotenv
python-jose[cryptography]
bcrypt